Use n-cluster (https://github.com/Financial-Times/n-cluster) to cluster across all cores
This is a basic implementation using the existing next solution.

This _could_ cause issues with constraints being created on startup, and will duplicate healthcheck work.

It's not overly straightforwards to remove the duplicate healthcheck work without either running it on the master and passing messages to the children, or running on the master and writing to disk.

This also moves a few devDependencies to dependencies which are being used in './server'
 🐿 v2.10.0