Fixes unlocking bug
When all fields are unlocked, lockedFields returns null. This resulted in it being impossible to unlock all fields because the previous logic for checking whether or not lockedFields had been modified was to check first if the value was null because null will never equal undefined and therefore no changes were saved to the db. This meant it was impossible to remove all fields. Now it checks if lockfFields and unlockFields query is used instead of whether or not lockedFields equals null. If all fields are unlocked it removes the property lockedFields without having to make any value changes.