Emits change event on start polling or on refresh for dev
Emits a change event when running api in dev when it starts polling or on refresh. Originally, when running the api locally with updateMode set to dev and schema linked, api was returning an undefined error because the the change callback was not being called which resulted in constructApi not being called.

This change creates a event change that is needed in order to construct the api.